src/ComplexNumberTest.java
import static org.junit.jupiter.api.Assertions.assertEquals;
import org.junit.jupiter.api.Test;

import java.util.ArrayList;

class ComplexNumberTest {

    @Test
    void whenInstantiatedInPolarForm_thenHasCorrectRectangularCoords() {
        ComplexNumber z1 = ComplexNumber.polar(1, 0);
        assertEquals(1, z1.realPart(), 0.01);
        assertEquals(0, z1.imaginaryPart(), 0.01);

        ComplexNumber z2 = ComplexNumber.polar(1, Math.PI / 4);
        assertEquals(Math.round(Math.sqrt(2) / 2), Math.round(z2.realPart()));
        assertEquals(Math.round(Math.sqrt(2) / 2), Math.round(z2.imaginaryPart()));

        ComplexNumber z3 = ComplexNumber.polar(1, 3 * Math.PI / 4);
        assertEquals(Math.round(-Math.sqrt(2) / 2), Math.round(z3.realPart()));
        assertEquals(Math.round(Math.sqrt(2) / 2), Math.round(z3.imaginaryPart()));

        ComplexNumber z4 = ComplexNumber.polar(1, 5 * Math.PI / 4);
        assertEquals(Math.round(-Math.sqrt(2) / 2), Math.round(z4.realPart()));
        assertEquals(Math.round(-Math.sqrt(2) / 2), Math.round(z4.imaginaryPart()));

        ComplexNumber z5 = ComplexNumber.polar(13, 5 * Math.PI / 4);
        assertEquals(Math.round(-13 * Math.sqrt(2) / 2), Math.round(z5.realPart()));
        assertEquals(Math.round(-13 * Math.sqrt(2) / 2), Math.round(z5.imaginaryPart()));

        ComplexNumber z6 = ComplexNumber.polar(1, 3 * (Math.PI / 2));
        assertEquals(0, Math.round(z6.realPart()));
        assertEquals(-1, z6.imaginaryPart(), 0.01);
    }
    @Test
    void testProductOfComplexNumbers() {
        var z1 = ComplexNumber.rectangular(1,2);
        var z2 = ComplexNumber.rectangular(2,3);
        var z3 = ComplexNumber.rectangular(3,4);
        assertEquals(-4,Math.round(ComplexNumber.product(z1,z2).realPart()));
        assertEquals(7, Math.round(ComplexNumber.product(z1,z2).imaginaryPart()));
        assertEquals(-40,Math.round(ComplexNumber.product(z1,z2,z3).realPart()));
        assertEquals(5, Math.round(ComplexNumber.product(z1,z2,z3).imaginaryPart()));
        assertEquals(-5,Math.round(ComplexNumber.product(z1,z3).realPart()));
        assertEquals(10, Math.round(ComplexNumber.product(z1,z3).imaginaryPart()));
        assertEquals(0, Math.round(ComplexNumber.product(z1,z2,ComplexNumber.ZERO,z3).realPart()));
        assertEquals(0, Math.round(ComplexNumber.product(z1,z2,ComplexNumber.ZERO,z3).imaginaryPart()));
    }
    @Test
    void testModulusOfComplexNumbers() {
        assertEquals(5, ComplexNumber.rectangular(3,4).modulus(), 0.001);
        assertEquals(10, ComplexNumber.polar(10, 0).modulus(), 0.001);
        assertEquals(39, ComplexNumber.rectangular(15,36).modulus(), 0.001);
        assertEquals(0, ComplexNumber.polar(0, 0).modulus(), 0.001);
        assertEquals(18*Math.sqrt(626)/5, ComplexNumber.rectangular(3.6, 90).modulus(), 0.001);
    }

    @Test
    void testSqrtOfComplexNumbers() {
        assert(ComplexNumber.rectangular(1,2).equals(ComplexNumber.rectangular(-3,4).sqrt()));
        assert(ComplexNumber.rectangular(0,3).equals(ComplexNumber.rectangular(-9,0).sqrt()));
        assert(ComplexNumber.rectangular(Math.sqrt(2),0)).equals((ComplexNumber.rectangular(2,0).sqrt()));
        assert(ComplexNumber.rectangular(8,-6).equals(ComplexNumber.rectangular(28,-96).sqrt()));
        assert(ComplexNumber.rectangular(24,63).equals(ComplexNumber.rectangular(-3393,3024).sqrt()));
        assert(ComplexNumber.rectangular(0.24,0.5).equals(ComplexNumber.rectangular(-0.1924,0.24).sqrt()));
        //Note tried a test with -2-7i but this will never happen when returning +ve sqrt
        assert(ComplexNumber.rectangular(2,7).equals(ComplexNumber.rectangular(-45,28).sqrt()));
    }

    @Test
    void testEquals(){
        ComplexNumber a = ComplexNumber.rectangular(3,4);
        assert(!ComplexNumber.rectangular(1,2).equals(a));
        ComplexNumber b = ComplexNumber.rectangular(4.6,1);
        assert(b.equals(ComplexNumber.rectangular(4.6,1)));
        assert(!b.equals(new ArrayList<String>()));

//        Problem with rounding when comparing - these 2 are identical but doesn't match due to way of implementation
//        System.out.println(a.toString() + ComplexNumber.polar(5, Math.atan(4.0/3.0)).toString());
//        assert(a.equals(ComplexNumber.polar(5, Math.atan(4.0/3.0))));
        assertEquals(3,Math.round(ComplexNumber.polar(5, Math.atan(4.0/3.0)).realPart()));
        assertEquals(4, Math.round(ComplexNumber.polar(5, Math.atan(4.0/3.0)).imaginaryPart()));
    }

    @Test
    void testSumOfComplexNumbers(){
        var z1 = ComplexNumber.rectangular(1,2);
        var z2 = ComplexNumber.rectangular(2,3);
        var z3 = ComplexNumber.rectangular(3,4);
        assert(ComplexNumber.rectangular(6,9).equals(ComplexNumber.sum(z1,z2,z3)));
        assert(ComplexNumber.rectangular(3,5).equals(ComplexNumber.sum(z1,z2)));
        assert(ComplexNumber.rectangular(4,6).equals(ComplexNumber.sum(z1,z3)));
        assert(ComplexNumber.rectangular(4,6).equals(ComplexNumber.sum(z1,z3,ComplexNumber.ZERO)));
        assert(ComplexNumber.ZERO.equals(ComplexNumber.sum(ComplexNumber.ZERO)));
        //TO DO more tests
    }
    @Test
    void testMultiplicativeInverse(){

    }

    @Test
    void whenInstantiatedInRectangularForm_thenHasCorrectArgument(){
        ComplexNumber z1 = ComplexNumber.rectangular(1, 1);
        assertEquals(Math.PI / 4, z1.argument().orElseThrow(), 0.001);

        ComplexNumber z2 = ComplexNumber.rectangular(-1, 1);
        assertEquals(3 * Math.PI / 4, z2.argument().orElse(-1.0), 0.001);

        ComplexNumber z3 = ComplexNumber.polar(1, 3 * Math.PI / 2);
        assertEquals(3 * Math.PI / 2, z3.argument().orElse(-1.0), 0.001);

        ComplexNumber z4 = ComplexNumber.rectangular(-3, 4);
        assertEquals(2.2143, z4.argument().orElse(-1.0), 0.001);

        assertEquals(5.878, ComplexNumber.rectangular(7,-3).argument().orElse(-1.0), 0.01);

        assertEquals(-1, ComplexNumber.rectangular(0,0).argument().orElse(-1.0));
    }


}